[MOV] mail: split alias model from its mixin
Purpose is to ease module understanding.

Task ID 2160070
